Extract operator value helpers
